test: create_solver_session
