+++ tests/test_codegen.py
@@ -5,7 +5,6 @@ import shutil
 import tempfile
 
 import pytest
-from util.solver_workflow import new_solver_session  # noqa: F401
 
 import ansys.fluent.core as pyfluent
 from ansys.fluent.core.codegen import StaticInfoType, allapigen
