+++ tests/test_error_handling.py
@@ -1,7 +1,6 @@
 import time
 
 import pytest
-from util.solver_workflow import new_solver_session  # noqa: F401
 
 
 @pytest.mark.fluent_version(">=23.1")
