+++ tests/test_field_data.py
@@ -1,6 +1,5 @@
 import numpy as np
 import pytest
-from util.solver_workflow import new_solver_session  # noqa: F401
 
 from ansys.fluent.core import examples
 from ansys.fluent.core.examples.downloads import download_file
