+++ tests/test_file_transfer_service.py
@@ -4,7 +4,6 @@ import os
 import pathlib
 
 import pytest
-from util.solver_workflow import new_solver_session  # noqa: F401
 
 from ansys.fluent.core import examples
 from ansys.fluent.core.utils.file_transfer_service import (
