+++ tests/test_fluent_fixes.py
@@ -1,5 +1,4 @@
 import pytest
-from util.solver_workflow import new_solver_session  # noqa: F401
 
 from ansys.fluent.core import examples
