+++ tests/test_fluent_session.py
@@ -6,10 +6,7 @@ import time
 import psutil
 import pytest
 from util.fixture_fluent import load_static_mixer_case  # noqa: F401
-from util.solver_workflow import (  # noqa: F401
-    new_solver_session,
-    new_solver_session_no_transcript,
-)
+from util.solver_workflow import new_solver_session_no_transcript  # noqa: F401
 
 import ansys.fluent.core as pyfluent
 from ansys.fluent.core.examples import download_file
