+++ tests/test_search.py
@@ -1,7 +1,6 @@
 import pytest
 from util.fixture_fluent import load_static_mixer_case  # noqa: F401
 from util.meshing_workflow import new_watertight_workflow_session  # noqa: F401
-from util.solver_workflow import new_solver_session  # noqa: F401
 
 import ansys.fluent.core as pyfluent
 from ansys.fluent.core.utils.search import (
