+++ tests/test_tui_api.py
@@ -1,5 +1,4 @@
 import pytest
-from util.solver_workflow import new_solver_session  # noqa: F401
 
 from ansys.fluent.core.services.datamodel_tui import TUIMenu
